application/requirements.txt
Flask==1.0.2
pytest==4.0.1
uwsgi==2.0.17
Flask-SQLAlchemy==2.3.2
PyMySQL==0.8.1
SQLAlchemy==1.2.14
pandas==0.23.3
redis==2.10.6
mysql-connector-python==8.0.15
